requirements.txt
streamlit
pandas
numpy
scikit-learn
nltk
xgboost
seaborn
matplotlib
google-api-python-client
python-dotenv